backend/requirements.txt
anyio==4.7.0
APScheduler==3.11.0
attrs==24.2.0
bitarray==3.0.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
click==8.1.7
Deprecated==1.2.15
flasgger==0.9.7.1
Flask[async]==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Limiter==3.9.2
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
limits==3.14.1
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.0.2
ordered-set==4.1.0
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg-pool==3.2.4
pycparser==2.22
Pygments==2.18.0
PyNaCl==1.5.0
python-dotenv==1.0.1
python-telegram-bot==21.9
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
tonsdk==1.0.15
typing_extensions==4.12.2
tzlocal==5.2
Werkzeug==3.1.3
wrapt==1.17.0
PyJWT==2.8.0
